+++ flmodules/src/template/mod.rs
@@ -3,4 +3,4 @@ pub mod core;
 // Messages for this module
 pub mod messages;
 // Integrating with other modules
-pub mod translation;
+pub mod broker;